@@ -5,7 +5,7 @@ version: 1.0.0+3
 homepage: https://github.com/dashdevs/alt_bloc
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
