@@ -12,14 +12,12 @@ typedef Consumer<T> = void Function(T state);
 /// [consumer] function will be triggered each time on new state.
 class BlocConsumer<B extends Bloc, S> extends BlocWidget<B, S> {
   const BlocConsumer(
-      {Key key,
-      @required this.child,
-      @required this.consumer,
-      B bloc,
-      Precondition<S> precondition})
-      : assert(child != null),
-        assert(consumer != null),
-        super(key: key, bloc: bloc, precondition: precondition);
+      {Key? key,
+      required this.child,
+      required this.consumer,
+      B? bloc,
+      Precondition<S>? precondition})
+      : super(key: key, bloc: bloc, precondition: precondition);
 
   final Widget child;
   final Consumer<S> consumer;
@@ -31,7 +29,7 @@ class BlocConsumer<B extends Bloc, S> extends BlocWidget<B, S> {
 class _BlocConsumerState<B extends Bloc, S>
     extends BlocWidgetState<B, S, BlocConsumer<B, S>> {
   @override
-  S get initialState => bloc.initialState<S>();
+  S? get initialState => bloc?.initialState<S>();
 
   @override
   Widget build(BuildContext context) => widget.child;
@@ -40,5 +38,5 @@ class _BlocConsumerState<B extends Bloc, S>
   void onNewState(S state) => widget.consumer.call(state);
 
   @override
-  Stream<S> get stream => bloc.getStateStream<S>();
+  Stream<S>? get stream => bloc?.getStateStream<S>();
 }
