@@ -3,9 +3,7 @@ import 'package:flutter/material.dart';
 
 import 'counter_bloc.dart';
 
-
 class CounterLayout extends StatelessWidget {
-
   CounterLayout({Key key, this.title}) : super(key: key);
 
   final String title;
@@ -26,29 +24,27 @@ class CounterLayout extends StatelessWidget {
                   'You have pushed the button this many times:',
                 ),
                 BlocBuilder<CounterBloc, int>(
-                    /*precondition: (prevCount, count) => count % 2 == 0*/
-                    builder: (_, count) {
-                      return Text(
-                        '$count',
-                        style: Theme.of(context).textTheme.display1,
-                      );
-                    }),
+                  /*precondition: (prevCount, count) => count % 2 == 0*/
+                  builder: (_, count) {
+                    return Text(
+                      '$count',
+                      style: Theme.of(context).textTheme.headline4,
+                    );
+                  },
+                ),
               ],
             ),
           ),
           Center(
             child: BlocBuilder<CounterBloc, bool>(
-                builder: (_, inProgress) {
-                  return inProgress ? CircularProgressIndicator() : Container();
-                }
+              builder: (_, inProgress) =>
+                  inProgress ? CircularProgressIndicator() : Container(),
             ),
-          )
+          ),
         ],
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: Provider
-            .of<CounterBloc>(context)
-            .increment,
+        onPressed: Provider.of<CounterBloc>(context).increment,
         tooltip: 'Increment',
         child: Icon(Icons.add),
       ),
