@@ -1,2 +1,2 @@
 /// Signature of predicate function that use to compare states of [Bloc].
-typedef Precondition<S> = bool Function(S previousState, S currentState);
+typedef Precondition<S> = bool Function(S? previousState, S? currentState);
